+++ src/libs/nav2_util/CMakeLists.txt
@@ -52,7 +52,7 @@ add_library(pf_lib SHARED
 add_library(sensors_lib SHARED
   src/sensors/amcl_sensor.cpp
   src/sensors/amcl_odom.cpp
-  src/sensors/amcl_laser.cpp
+  src/sensors/laser.cpp
 )
 
 add_library(map_loader SHARED
