+++ src/localization/nav2_amcl/src/amcl_node.hpp
@@ -35,7 +35,7 @@
 #include "std_srvs/srv/empty.hpp"
 #include "tf2_ros/transform_broadcaster.h"
 #include "tf2_ros/transform_listener.h"
-#include "nav2_util/sensors/amcl_laser.h"
+#include "nav2_util/sensors/laser.h"
 #include "nav2_util/sensors/amcl_odom.h"
 
 #define NEW_UNIFORM_SAMPLING 1
